Register commands explicitly in dedicated command registries
Allow explicit command registries to be provided by projects offering an extensible CLI as part of their offerings.

Move registration outside of commands so it can be done against explicitly named command registries.

This arrangement makes it possible for multiple, distinct sets of commands to be registered for different users of hanami-cli, all within the same Ruby runtime environment.

To support this broad change, a couple of other things have changed too:

- `Hanami::Cli::Command` is now a class, and commands should subclass it.
- A class method interface has been created for the Command class’ `desc` and `params`, with those methods setting class-level instance variables.
- `Hanami::Cli` is now a class, and it should be instantiated with a command registry object
- `Hanami::Cli::Renderer` is now instantiated with a command registry object, and `#call`-ede with the name of the command