Instantiate commands with options and provide command API via instances
This makes it possible for dependencies or other sorts of configuration to be passed to commands at registration time.

The command_name is also passed to the command as it is instantiated, so we avoid mutating class-level ivars across potentially multiple registrations of the same command class.

Also, by working with command instances throughout the framework (i.e. accessing command description, examples, arguments, etc. from the instance, rather than the class), it means that we better support any sort of object willing to provide the command API, rather than just Hanami::Cli::Command subclasses.