Remove support for plain callable objects as registered commands
These lack all the affordances that make working with the CLI nice (such as descriptions, argument list, other docs, etc.).

Instead of supporting plain callable objects, we should instead expect objects that conform to behaviors of Hanami::Cli::Command. Now that these are worked with as instances throughout the framework (rather than expecting a mix of instance- and class-level methods), this is a much more reasonable expectation to make.