Do not report errors for commands with subcommands
This commit fixes the issue where commands with subcommands, when
executed with --help, would report the main command as an error.

This commit also refactors handling of the unknown command into
it's own method, and additionally fixes the issue where an ARGV
containing -- may trigger unknown command error with arguments
that follow --.

Finally, adding a couple of new specs to verify that subcommands
are working as expected.