Use open3 in inherited_command_spec.rb (#153)
`based --help` results in putting text both in stdout and stderr.
However, it seems that they might be flushed in different order between
JRuby and MRI (or maybe just in JRuby's implementation of backticks). As
a result, the test using backticks was failing on whitespace.

However, this (whitespace) is not really a very important part of the
test. When we switch testing to Open3, like for example
spell_checker_spec.rb does, we get stable results between both
implementations.